badger: try and optimize query function
