+++ badger/datastore.go
@@ -1,7 +1,6 @@
 package dsbadger
 
 import (
-	"context"
 	"errors"
 	"log"
 	"time"
@@ -517,169 +516,6 @@ func (t *txn) Query(q dsq.Query) (dsq.Results, error) {
 		return nil, ErrClosed
 	}
 	return t.query(q)
-}
-
-func (t *txn) query(q dsq.Query) (dsq.Results, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	opt := badger.DefaultIteratorOptions
-	opt.PrefetchValues = !q.KeysOnly
-	prefix := ds.NewKey(q.Prefix).String()
-	if prefix != "/" {
-		opt.Prefix = []byte(prefix + "/")
-	}
-	// Handle ordering
-	if len(q.Orders) > 0 {
-		switch q.Orders[0].(type) {
-		case dsq.OrderByKey, *dsq.OrderByKey:
-		// We order by key by default.
-		case dsq.OrderByKeyDescending, *dsq.OrderByKeyDescending:
-			// Reverse order by key
-			opt.Reverse = true
-		default:
-			// Ok, we have a weird order we can't handle. Let's
-			// perform the _base_ query (prefix, filter, etc.), then
-			// handle sort/offset/limit later.
-
-			// Skip the stuff we can't apply.
-			baseQuery := q
-			baseQuery.Limit = 0
-			baseQuery.Offset = 0
-			baseQuery.Orders = nil
-
-			// perform the base query.
-			res, err := t.query(baseQuery)
-			if err != nil {
-				return nil, err
-			}
-
-			// fix the query
-			res = dsq.ResultsReplaceQuery(res, q)
-
-			// Remove the parts we've already applied.
-			naiveQuery := q
-			naiveQuery.Prefix = ""
-			naiveQuery.Filters = nil
-
-			// Apply the rest of the query
-			return dsq.NaiveQueryApply(naiveQuery, res), nil
-		}
-	}
-	var (
-		it         = t.txn.NewIterator(opt)
-		done       = make(chan bool)
-		resultChan = make(chan dsq.Result)
-		entries    = make([]dsq.Entry, 0)
-	)
-	go func() {
-		defer func() {
-			done <- true
-		}()
-		if t.ds.closed.IsSet() {
-			return
-		}
-		// this iterator is part of an implicit transaction, so when
-		// we're done we must discard the transaction. It's safe to
-		// discard the txn it because it contains the iterator only.
-		if t.implicit {
-			defer t.discard()
-		}
-		defer it.Close()
-		// All iterators must be started by rewinding.
-		it.Rewind()
-		// skip to the offset
-		for skipped := 0; skipped < q.Offset && it.Valid(); it.Next() {
-			// On the happy path, we have no filters and we can go
-			// on our way.
-			if len(q.Filters) == 0 {
-				skipped++
-				continue
-			}
-			// On the sad path, we need to apply filters before
-			// counting the item as "skipped" as the offset comes
-			// _after_ the filter.
-			item := it.Item()
-			matches := true
-			check := func(value []byte) error {
-				e := dsq.Entry{Key: string(item.Key()), Value: value, Size: int(item.ValueSize())}
-				// Only calculate expirations if we need them.
-				if q.ReturnExpirations {
-					e.Expiration = expires(item)
-				}
-				matches = filter(q.Filters, e)
-				return nil
-			}
-			// Maybe check with the value, only if we need it.
-			var err error
-			if q.KeysOnly {
-				err = check(nil)
-			} else {
-				err = item.Value(check)
-			}
-			if err != nil {
-				select {
-				case resultChan <- dsq.Result{Error: err}:
-				case <-t.ds.closing:
-					return
-				case <-ctx.Done():
-					return
-				}
-			}
-			if !matches {
-				skipped++
-			}
-		}
-		for sent := 0; (q.Limit <= 0 || sent < q.Limit) && it.Valid(); it.Next() {
-			item := it.Item()
-			e := dsq.Entry{Key: string(item.Key())}
-			// Maybe get the value
-			var result dsq.Result
-			if !q.KeysOnly {
-				b, err := item.ValueCopy(nil)
-				if err != nil {
-					result = dsq.Result{Error: err}
-				} else {
-					e.Value = b
-					e.Size = len(b)
-					result = dsq.Result{Entry: e}
-				}
-			} else {
-				e.Size = int(item.ValueSize())
-				result = dsq.Result{Entry: e}
-			}
-
-			if q.ReturnExpirations {
-				result.Expiration = expires(item)
-			}
-
-			// Finally, filter it (unless we're dealing with an error).
-			if result.Error == nil && filter(q.Filters, e) {
-				continue
-			}
-
-			select {
-			case resultChan <- result:
-				sent++
-			case <-ctx.Done():
-				return
-			case <-t.ds.closing:
-				return
-			}
-		}
-	}()
-	for {
-		select {
-		case <-done:
-			goto FINISHED
-		case result := <-resultChan:
-			if result.Error != nil {
-				log.Println("query result failure: ", result.Error)
-			}
-			entries = append(entries, result.Entry)
-		}
-	}
-FINISHED:
-	return dsq.ResultsWithEntries(q, entries), nil
 }
 
 func (t *txn) Commit() error {
